Move generator code into compiledb __init__.py
