+++ compiledb/__main__.py
@@ -19,9 +19,9 @@
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-from compiledb.generator import _main
+from . import cli
 
 if(__name__ == "__main__"):
-    _main()
+    cli()
 
 # ex: ts=2 sw=4 et filetype=python
